+++ lib/ReactViewModels/ViewState.ts
@@ -10,7 +10,11 @@ import { Ref } from "react";
 import defined from "terriajs-cesium/Source/Core/defined";
 import CesiumEvent from "terriajs-cesium/Source/Core/Event";
 import addedByUser from "../Core/addedByUser";
-import { Category, HelpAction } from "../Core/AnalyticEvents/analyticEvents";
+import {
+  Category,
+  HelpAction,
+  StoryAction
+} from "../Core/AnalyticEvents/analyticEvents";
 import Result from "../Core/Result";
 import triggerResize from "../Core/triggerResize";
 import PickedFeatures from "../Map/PickedFeatures";
@@ -23,6 +27,7 @@ import { BaseModel } from "../Models/Definition/Model";
 import getAncestors from "../Models/getAncestors";
 import Terria from "../Models/Terria";
 import { SATELLITE_HELP_PROMPT_KEY } from "../ReactViews/HelpScreens/SatelliteHelpPrompt";
+import { animationDuration } from "../ReactViews/StandardUserInterface/StandardUserInterface";
 import {
   defaultTourPoints,
   RelativePosition,
@@ -95,6 +100,9 @@ export default class ViewState {
   @observable selectedTrainerItem: string = "";
   @observable currentTrainerItemIndex: number = 0;
   @observable currentTrainerStepIndex: number = 0;
+
+  @observable printWindow: Window | null = null;
+
   @action
   setSelectedTrainerItem(trainerItem: string) {
     this.selectedTrainerItem = trainerItem;
@@ -545,13 +553,25 @@ export default class ViewState {
     openAddData = true
   ): Promise<Result<void>> {
     try {
+      // Get referenced target first.
+      if (ReferenceMixin.isMixedInto(item)) {
+        (await item.loadReference()).throwIfError();
+        if (item.target) {
+          return this.viewCatalogMember(item.target);
+        } else {
+          return Result.error(`Could not view catalog member ${getName(item)}`);
+        }
+      }
+      const theItem: BaseModel =
+        ReferenceMixin.isMixedInto(item) && item.target ? item.target : item;
+
       // Set preview item
-      runInAction(() => (this._previewedItem = item));
+      runInAction(() => (this._previewedItem = theItem));
 
       // Open "Add Data"
       if (openAddData) {
-        if (addedByUser(item)) {
-          runInAction(() => (this.userDataPreviewedItem = item));
+        if (addedByUser(theItem)) {
+          runInAction(() => (this.userDataPreviewedItem = theItem));
 
           this.openUserData();
         } else {
@@ -567,32 +587,21 @@ export default class ViewState {
           });
         }
 
-        // mobile switch to nowvewing if not viewing a group
-        if (!GroupMixin.isMixedInto(item)) {
+        // mobile switch to now vewing if not viewing a group
+        if (!GroupMixin.isMixedInto(theItem)) {
           this.switchMobileView(this.mobileViewOptions.preview);
         }
       }
 
-      // Load preview item
-      if (ReferenceMixin.isMixedInto(item)) {
-        (await item.loadReference()).throwIfError();
-
-        // call viewCatalogMember on reference.target
-        if (item.target) {
-          return await this.viewCatalogMember(item.target, isOpen, stratum);
+      if (GroupMixin.isMixedInto(theItem)) {
+        theItem.setTrait(stratum, "isOpen", isOpen);
+        if (theItem.isOpen) {
+          (await theItem.loadMembers()).throwIfError();
         }
-        return Result.error(`Failed to resolve reference for ${getName(item)}`);
-      }
-
-      if (GroupMixin.isMixedInto(item)) {
-        item.setTrait(stratum, "isOpen", isOpen);
-        if (item.isOpen) {
-          (await item.loadMembers()).throwIfError();
-        }
-      } else if (MappableMixin.isMixedInto(item))
-        (await item.loadMapItems()).throwIfError();
-      else if (CatalogMemberMixin.isMixedInto(item))
-        (await item.loadMetadata()).throwIfError();
+      } else if (MappableMixin.isMixedInto(theItem))
+        (await theItem.loadMapItems()).throwIfError();
+      else if (CatalogMemberMixin.isMixedInto(theItem))
+        (await theItem.loadMetadata()).throwIfError();
     } catch (e) {
       return Result.error(e, `Could not view catalog member ${getName(item)}`);
     }
@@ -705,10 +714,31 @@ export default class ViewState {
     this.currentTool = undefined;
   }
 
+  @action setPrintWindow(window: Window | null) {
+    if (this.printWindow) {
+      this.printWindow.close();
+    }
+    this.printWindow = window;
+  }
+
   @action
   toggleMobileMenu() {
     this.setTopElement("mobileMenu");
     this.mobileMenuVisible = !this.mobileMenuVisible;
+  }
+
+  @action
+  runStories() {
+    this.storyBuilderShown = false;
+    this.storyShown = true;
+
+    setTimeout(function() {
+      triggerResize();
+    }, animationDuration || 1);
+
+    this.terria.currentViewer.notifyRepaintRequired();
+
+    this.terria.analytics?.logEvent(Category.story, StoryAction.runStory);
   }
 
   @computed