+++ lib/ReactViews/Workbench/Controls/ViewingControls.jsx
@@ -36,6 +36,7 @@ import SplitterTraits from "../../../Traits/TraitsClasses/SplitterTraits";
 import { exportData } from "../../Preview/ExportData";
 import LazyItemSearchTool from "../../Tools/ItemSearchTool/LazyItemSearchTool";
 import WorkbenchButton from "../WorkbenchButton";
+import CameraView from "../../../Models/CameraView";
 
 const BoxViewingControl = styled(Box).attrs({
   centered: true,
@@ -129,12 +130,31 @@ const ViewingControls = observer(
         const item = this.props.item;
         let zoomToView = item;
         if (
+          item.idealZoom !== undefined &&
+          item.idealZoom.targetLongitude !== undefined &&
+          item.idealZoom.targetLatitude !== undefined &&
+          item.idealZoom.range >= 0
+        ) {
+          // No value checking here. Improper values can lead to unexpected results.
+          const lookAt = {
+            targetLongitude: item.idealZoom.targetLongitude,
+            targetLatitude: item.idealZoom.targetLatitude,
+            targetHeight: item.idealZoom.targetHeight,
+            heading: item.idealZoom.heading,
+            pitch: item.idealZoom.pitch,
+            range: item.idealZoom.range
+          };
+
+          // In the case of 2D viewer, it zooms to rectangle area approximated by the camera view parameters.
+          zoomToView = CameraView.fromJson({ lookAt: lookAt });
+        } else if (
           item.rectangle !== undefined &&
           item.rectangle.east - item.rectangle.west >= 360
         ) {
           zoomToView = this.props.viewState.terria.mainViewer.homeCamera;
           console.log("Extent is wider than world so using homeCamera.");
         }
+
         this.setState({ isMapZoomingToCatalogItem: true });
         viewer.zoomTo(zoomToView).finally(() => {
           this.setState({ isMapZoomingToCatalogItem: false });