+++ lib/Traits/TraitsClasses/WebMapServiceCatalogGroupTraits.ts
@@ -1,12 +1,14 @@
 import { JsonObject } from "../../Core/Json";
 import anyTrait from "../Decorators/anyTrait";
+import objectTrait from "../Decorators/objectTrait";
+import primitiveTrait from "../Decorators/primitiveTrait";
+import mixTraits from "../mixTraits";
 import CatalogMemberTraits from "./CatalogMemberTraits";
+import ExportWebCoverageServiceTraits from "./ExportWebCoverageServiceTraits";
 import GetCapabilitiesTraits from "./GetCapabilitiesTraits";
 import GroupTraits from "./GroupTraits";
-import mixTraits from "../mixTraits";
-import primitiveTrait from "../Decorators/primitiveTrait";
-import UrlTraits from "./UrlTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
+import UrlTraits from "./UrlTraits";
 
 export default class WebMapServiceCatalogGroupTraits extends mixTraits(
   GetCapabilitiesTraits,
@@ -28,4 +30,12 @@ export default class WebMapServiceCatalogGroupTraits extends mixTraits(
     description: "Sets traits on child WebMapServiceCatalogItem's"
   })
   itemProperties?: JsonObject;
+
+  @objectTrait({
+    name: "Per layer WebCoverageService",
+    description:
+      "To enable Export/WebCoverageService for **all** WMS layers in this group, set `perLayerLinkedWcs.linkedWcsUrl`. `linkedWcsCoverage` will be set to the WMS layer `Name` if it is defined, layer `Title` otherwise.",
+    type: ExportWebCoverageServiceTraits
+  })
+  perLayerLinkedWcs?: ExportWebCoverageServiceTraits;
 }